requirements.txt
asgiref==3.6.0
attrs==23.1.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
Faker==19.6.2
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.5.0
PyJWT==1.7.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
